\docType{class}
\name{sum.bma-class}
\alias{initialize,}
\alias{print,}
\alias{show,}
\alias{sum,}
\alias{sum.bma-class}
\alias{sum.bma-method}
\title{An object of '\code{sum.bma}' class.}
\description{
An Object of class '\code{sum.bma}' is created when the
\code{summary} function is executed on an object of class
'\code{bma}'.
}
\details{
An object of the class '\code{sum.bma}' has the following
slots: \itemize{ \item{expected.coefficient}{The expected
value of each coefficient} \item{Posterior.prob}{The
posterior probability that the coefficient is non-zero} }
}
\author{
Myunghoon Kang: \email{myunghoon@wustl.edu}
}

