\docType{class}
\name{bma-class}
\alias{bma-class}
\title{A bma class}
\description{
Object of class \code{bma} as created by the \code{fitBMA}
functions
}
\details{
An object of the class `bma' has the following slots:
\itemize{ \item \code{combo.coef} A list of coefficients
for each possible regression \item \code{combo.fit} A
vector of R-Squared values \item \code{bmk} \item
\code{exp.vals} \item \code{thecoefs} \item \code{theses}
\item \code{coefprobs} \item
\code{coefprobs.largerthanzero} \item
\code{conditional.sds} \item \code{x} A matrix input of
covariates \item \code{y} The dependent variable input
(vector) }
}
\author{
Jacob Montgomery, Dino Hadzic, Jae Hee Jung, Emily Moore
}

