\docType{class}
\name{BMA-class}
\alias{BMA-class}
\alias{initialize,BMA-method}
\title{Bayesian Model Averaging}
\description{
Object of class \code{BMA} is created by \code{fitBMA},
\code{plot}, and \code{return} functions
}
\details{
An object of class 'BMA' has the following slots: \itemize{
\item \code{Coefs} Coefficient estimates \item \code{R2s}
Rsquared values \item \code{ModelOdds} Posterior model odds
\item \code{ExpectedValues} Posterior expected values of
coefficients \item \code{Nonzero} Posterior probability
that the coefficient is non-zero }
}
\author{
Jae Hee Jung: \email{jaeheejung@wustl.edu}
}

