\name{fitBMA}
\alias{fitBMA}
\alias{fitBMA,ANY-method}
\title{Performs Bayesian Model Averaging (BMA) over entire model space}
\usage{
fitBMA(M, Y, gprior)
}
\arguments{
  \item{M}{a matrix of observations for the covariates.}

  \item{Y}{a numeric vector whose length is equal to the
  number of rows of \code{M}.}

  \item{gprior}{a user specified prior.}
}
\value{
An object of class list with elements \item{Coeffs}{A
matrix of coefficient estimates generated by the
regressions of all possible covariate combinations.}
\item{R2s}{A vector of R-squared values generated by the
regressions of all possible covariate combinations.}
\item{PosteriorOdds}{The posterior model odds for each
model.} \item{PosteriorValues}{The posterior expected value
of each coefficient.} \item{PosteriorNonZero}{The posterior
probability that the coefficient is non-zero.}
}
\description{
Performs BMA over entire model space and produces relevant
output.
}
\examples{
M <- matrix(rnorm(30), ncol=3)
 Y <- c(rnorm(10))
 fitBMA(M, Y, gprior=3)
}
\author{
Dino Hadzic \email{dino.hadzic@wustl.edu}
}
\seealso{
\code{\link{plotBMA}}
}

